moved theme to its own file, added comment to pages/index.js
